wip: renaming test definition files to have ee suffix
