+++ packages/cli/src/databases/entities/index.ts
@@ -20,7 +20,7 @@ import { Settings } from './settings';
 import { SharedCredentials } from './shared-credentials';
 import { SharedWorkflow } from './shared-workflow';
 import { TagEntity } from './tag-entity';
-import { TestDefinition } from './test-definition';
+import { TestDefinition } from './test-definition.ee';
 import { User } from './user';
 import { Variables } from './variables';
 import { WebhookEntity } from './webhook-entity';
