+++ packages/cli/src/generic-helpers.ts
@@ -3,7 +3,7 @@ import { validate } from 'class-validator';
 import type { AnnotationTagEntity } from '@/databases/entities/annotation-tag-entity.ee';
 import type { CredentialsEntity } from '@/databases/entities/credentials-entity';
 import type { TagEntity } from '@/databases/entities/tag-entity';
-import type { TestDefinition } from '@/databases/entities/test-definition';
+import type { TestDefinition } from '@/databases/entities/test-definition.ee';
 import type { User } from '@/databases/entities/user';
 import type { WorkflowEntity } from '@/databases/entities/workflow-entity';
